fix(security): patch path traversal, scoring bug, race condition, and code quality issues
- Add safe_resolve() containment check on all user-controlled file paths
  (manual resources, model pickle loads, signal reads) to prevent path traversal
- Guard fault_signal_files=None in train_anomaly_model to prevent TypeError crash
- Fix bearing confidence scoring: 1 harmonic now returns "moderate" (was "high")
- Thread-safe singleton for SignalRepository with double-checked locking
- Sanitize feature output filename to prevent directory escape
- Validate model_name before constructing pickle paths
- Extract shared load_and_validate_metadata to _utils.py (eliminate 115-line duplication)
- Use np.random.default_rng() in generate_test_signal for reproducibility
- Raise coverage fail_under from 60% to 85%

All 414 tests pass, coverage 86.95%.
