+++ pyproject.toml
@@ -181,7 +181,7 @@ omit = [
 ]
 
 [tool.coverage.report]
-fail_under = 60
+fail_under = 85
 exclude_lines = [
     "pragma: no cover",
     "def __repr__",
